removed keeper SetConversionRate and SetFreezePeriod and related messages, tx, etc.
