+++ x/commerciomint/types/msgs_test.go
@@ -70,52 +70,3 @@ func TestMsgBurnCCC_ValidateBasic(t *testing.T) {
 	require.Error(t, NewMsgBurnCCC(testOwner, "", sdk.NewCoin("uccc", sdk.NewInt(100))).ValidateBasic())
 	require.NoError(t, NewMsgBurnCCC(testOwner, uuid, sdk.NewCoin("uccc", sdk.NewInt(100))).ValidateBasic())
 }
-
-func TestMsgSetCCCConversionRate_ValidateBasic(t *testing.T) {
-
-	tests := []struct {
-		name           string
-		signer         sdk.AccAddress
-		collateralRate sdk.Dec
-		wantErr        bool
-	}{
-		{"empty signer", nil, sdk.NewDec(2), true},
-		{"invalid signer", []byte("test"), sdk.NewDec(2), true},
-		{"ok", testOwner, sdk.NewDec(2), false},
-		{"zero collateral rate", []byte("test"), sdk.NewDec(0), true},
-		{"negative collateral rate", []byte("test"), sdk.NewDec(-1), true},
-	}
-	for _, tt := range tests {
-		tt := tt
-		t.Run(tt.name, func(t *testing.T) {
-			msg := NewMsgSetCCCConversionRate(tt.signer, tt.collateralRate)
-			require.Equal(t, "commerciomint", msg.Route())
-			require.Equal(t, "setEtpsConversionRate", msg.Type())
-			require.Equal(t, tt.wantErr, msg.ValidateBasic() != nil)
-		})
-	}
-}
-
-func TestMsgSetCCCFreezePeriod_ValidateBasic(t *testing.T) {
-
-	tests := []struct {
-		name         string
-		signer       sdk.AccAddress
-		freezePeriod string
-		wantErr      bool
-	}{
-		{"invalid signer", []byte("test"), "1m", true},
-		{"negative duration", []byte("test"), "-1m", true},
-		{"ok", testOwner, "1m", false},
-		{"negative duration", testOwner, "A", true},
-	}
-	for _, tt := range tests {
-		tt := tt
-		t.Run(tt.name, func(t *testing.T) {
-			msg := NewMsgSetCCCFreezePeriod(tt.signer, tt.freezePeriod)
-			require.Equal(t, "commerciomint", msg.Route())
-			require.Equal(t, "setEtpsFreezePeriod", msg.Type())
-			require.Equal(t, tt.wantErr, msg.ValidateBasic() != nil)
-		})
-	}
-}
